[task 5] Delete old prompt modal, move shared types to newsession.go and modalWidth to feedbacknotemodal.go
